+++ packages/screenshoter/test/e2e/ios.spec.js
@@ -20,14 +20,14 @@ const env = {
     accessKey: process.env.SAUCE_ACCESS_KEY,
   },
 
-  // url: 'http://0.0.0.0:4723/wd/hub',
-  // capabilities: {
-  //   deviceName: 'iPhone 11 Pro',
-  //   platformName: 'iOS',
-  //   platformVersion: '14.5',
-  //   automationName: 'XCUITest',
-  //   app: 'https://applitools.jfrog.io/artifactory/Examples/IOSTestApp/1.9/app/IOSTestApp.zip',
-  // },
+  url: 'http://0.0.0.0:4723/wd/hub',
+  capabilities: {
+    deviceName: 'iPhone 11 Pro',
+    platformName: 'iOS',
+    platformVersion: '14.5',
+    automationName: 'XCUITest',
+    app: 'https://applitools.jfrog.io/artifactory/Examples/IOSTestApp/1.9/app/IOSTestApp.zip',
+  },
 }
 
 describe('screenshoter ios', () => {
@@ -103,8 +103,8 @@ describe('screenshoter ios', () => {
     return fullApp({type: 'superview'})
   })
 
-  it('take webview screenshot', () => {
-    return webview()
+  it.only('take webview screenshot', () => {
+    return webview({debug: {path: './'}})
   })
 
   it('take full webview screenshot', () => {
@@ -203,6 +203,7 @@ describe('screenshoter ios', () => {
     await driver.target.getContexts()
     await utils.general.sleep(500)
     const [, webview] = await driver.target.getContexts()
+    console.log(webview)
     await driver.target.switchContext(webview)
 
     await driver.init()