+++ packages/screenshoter/test/e2e/web-ios.spec.js
@@ -18,6 +18,19 @@ const env = {
     username: process.env.SAUCE_USERNAME,
     accessKey: process.env.SAUCE_ACCESS_KEY,
   },
+
+  // url: 'http://0.0.0.0:4723/wd/hub',
+  // capabilities: {
+  //   name: 'iOS Web Screenshoter Test',
+  //   deviceName: 'iPhone 11 Pro',
+  //   browserName: 'safari',
+  //   platformName: 'iOS',
+  //   platformVersion: '14.5',
+  //   appiumVersion: '1.20.1',
+  //   automationName: 'XCUITest',
+  //   username: process.env.SAUCE_USERNAME,
+  //   accessKey: process.env.SAUCE_ACCESS_KEY,
+  // },
 }
 
 describe('screenshoter web ios', () => {
@@ -39,18 +52,33 @@ describe('screenshoter web ios', () => {
   })
 
   it('take viewport screenshot', () => {
-    return viewport()
+    return viewport({orientation: 'portrait'})
+  })
+
+  it('take viewport screenshot with landscape orientation', () => {
+    return viewport({orientation: 'landscape'})
   })
 
   it('take full page screenshot', () => {
-    return fullPage()
+    return fullPage({orientation: 'portrait'})
   })
 
-  async function viewport(options) {
+  it('take full page screenshot with landscape orientation', () => {
+    return fullPage({orientation: 'landscape'})
+  })
+
+  async function viewport({orientation = 'portrait', ...options} = {}) {
+    const expectedPath = `./test/fixtures/web-ios/page${orientation === 'landscape' ? '-landscape' : ''}.png`
+
+    await driver.target.setOrientation(orientation.toUpperCase())
+    await driver.visit(await driver.getUrl())
+
+    await driver.init()
+
     const screenshot = await takeScreenshot({logger, driver, ...options})
     try {
       const actual = await screenshot.image.toObject()
-      const expected = await makeImage('./test/fixtures/web-ios/page.png').toObject()
+      const expected = await makeImage(expectedPath).toObject()
       assert.strictEqual(pixelmatch(actual.data, expected.data, null, expected.width, expected.height), 0)
     } catch (err) {
       await screenshot.image.debug({path: './logs', name: 'ios_viewport_failed'})
@@ -58,11 +86,18 @@ describe('screenshoter web ios', () => {
     }
   }
 
-  async function fullPage(options) {
+  async function fullPage({orientation = 'portrait', ...options} = {}) {
+    const expectedPath = `./test/fixtures/web-ios/page-fully${orientation === 'landscape' ? '-landscape' : ''}.png`
+
+    await driver.target.setOrientation(orientation.toUpperCase())
+    await driver.visit(await driver.getUrl())
+
+    await driver.init()
+
     const screenshot = await takeScreenshot({logger, driver, fully: true, ...options})
     try {
       const actual = await screenshot.image.toObject()
-      const expected = await makeImage('./test/fixtures/web-ios/page-fully.png').toObject()
+      const expected = await makeImage(expectedPath).toObject()
       assert.strictEqual(pixelmatch(actual.data, expected.data, null, expected.width, expected.height), 0)
     } catch (err) {
       await screenshot.image.debug({path: './logs', name: 'ios_full_page_failed'})